requirements.txt
Flask==3.1.2
tensorflow==2.20.0
numpy==2.2.6
pillow==12.0.0
rembg==2.0.67
Werkzeug==3.1.3
